+++ udata_front/theme/gouvfr/assets/js/.storybook/preview.ts
@@ -7,6 +7,7 @@ import i18n from "../i18n";
 import Api from "../plugins/api";
 import Container from "../components/Ui/Container/Container.vue";
 import Datasets from "../pages/Admin/Datasets/Datasets.vue";
+import Discussions from "../pages/Admin/Discussions/Discussions.vue";
 import Members from "../pages/Admin/Members/Members.vue";
 import Profile from "../pages/Admin/Profile/Profile.vue";
 import Reuses from "../pages/Admin/Reuses/Reuses.vue";
@@ -26,6 +27,7 @@ setup((app) => {
       children: [
         { path: 'datasets', component: Datasets, props: true, name: "organization-datasets" },
         { path: 'reuses', component: Reuses, props: true, name: "organization-reuses" },
+        { path: 'discussions', component: Discussions, props: true, name: "organization-discussions" },
         { path: 'members', component: Members, props: true, name: "organization-members" },
         { path: 'profile', component: Profile, props: true, name: "organization-profile" },
       ],